sl5 build warning removed
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@1392 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ arich/hitarich/include/ARICHSimHit.h
@@ -93,7 +93,7 @@ namespace Belle2 {
     TVector3 m_position;      /*!< Local position of hit (in module coordinates) */
     double m_globalTime;      /*!< Global time of hit */
     double m_energy;          /*!< Energy of detected photon */
-    double m_parentID;        /*!< G4 ID number of photons parent particle */
+    int m_parentID;        /*!< G4 ID number of photons parent particle */
 
 
     ClassDef(ARICHSimHit, 1); /*!< the class title */
